betting-research/requirements.txt
# Kalshi API
requests>=2.31.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Data processing
pandas>=2.1.4
numpy>=1.26.2

# Machine Learning
scikit-learn>=1.3.2

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
